Fix penalty rounding in the standing
* Fix the rounding of the penalty to match the ICPC rules, i.e. times per accepted problem must be first rounded to minutes and then added up.
* Show the penalty in ICPC-style, i.e. show the total number of minutes

This was compared with the Saris' standing and the results match.

+++ tests/test_standing_calculation.py
@@ -220,7 +220,7 @@ class CalculateStandingCase(FixturedTestCase):
 
             self.assertIsInstance(rank_entry_new.solved, int)
             self.assertIsInstance(rank_entry_new.rank, int)
-            self.assertIsInstance(rank_entry_new.penalty, float)
+            self.assertIsInstance(rank_entry_new.penalty, int)
             self.assertIsInstance(rank_entry_new.participant, ContestInstance)
 
             if debug:
